+++ internal/app/grafana-multi-tenant-proxy/handler/auth/auth.go
@@ -6,8 +6,9 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/giantswarm/grafana-multi-tenant-proxy/internal/pkg"
 	"go.uber.org/zap"
+
+	"github.com/giantswarm/grafana-multi-tenant-proxy/internal/app/grafana-multi-tenant-proxy/config"
 )
 
 type key int
@@ -19,20 +20,21 @@ const (
 
 // INTERFACE to handle different type of authentication
 type Authenticator interface {
-	Authenticate(r *http.Request) (bool, string)
+	Authenticate(r *http.Request, targetServer *config.TargetServer) (bool, string)
 	OnAuthenticationError(w http.ResponseWriter)
 }
+
 type AuthenticationMiddleware struct {
-	handler    http.HandlerFunc
-	authConfig *pkg.Authn
-	logger     *zap.Logger
+	handler http.HandlerFunc
+	config  *config.Config
+	logger  *zap.Logger
 }
 
-func NewAuthenticationMiddleware(logger *zap.Logger, handler http.HandlerFunc, authConfig pkg.Authn) *AuthenticationMiddleware {
+func NewAuthenticationMiddleware(config *config.Config, logger *zap.Logger, handler http.HandlerFunc) *AuthenticationMiddleware {
 	return &AuthenticationMiddleware{
-		handler:    handler,
-		authConfig: &authConfig,
-		logger:     logger,
+		handler: handler,
+		config:  config,
+		logger:  logger,
 	}
 }
 
@@ -40,15 +42,28 @@ func NewAuthenticationMiddleware(logger *zap.Logger, handler http.HandlerFunc, a
 // Authenticate can be used as a middleware chain to authenticate every request before proxying the request
 func (am AuthenticationMiddleware) Authenticate() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		authenticator, err := newAuthenticator(r, am.authConfig, am.logger)
+		authenticator, err := newAuthenticator(r, am.config, am.logger)
 		if err != nil {
-			am.logger.Error(fmt.Sprintf("Error while authenticating request %s", r.URL), zap.Error(err))
+			am.logger.Error("Error while authenticating request", zap.String("url", r.URL.String()), zap.Error(err))
 			w.WriteHeader(401)
 			w.Write([]byte("Unauthorised\n"))
 			return
 		}
+
+		targetServer := am.config.Proxy.FindTargetServer(r.Host)
+		if targetServer == nil {
+			am.logger.Error("Target server not configured",
+				zap.String("host", r.Host),
+				zap.String("url", r.URL.String()),
+				zap.Error(err),
+			)
+			w.WriteHeader(404)
+			w.Write([]byte("Not found\n"))
+			return
+		}
+
 		am.logger.Debug(fmt.Sprintf("Authentication mode: %T", authenticator))
-		ok, orgID := authenticator.Authenticate(r)
+		ok, orgID := authenticator.Authenticate(r, targetServer)
 		if !ok {
 			authenticator.OnAuthenticationError(w)
 			return
@@ -58,30 +73,29 @@ func (am AuthenticationMiddleware) Authenticate() http.HandlerFunc {
 	}
 }
 
-func (am AuthenticationMiddleware) ApplyConfig(authConfig pkg.Authn) {
-	*am.authConfig = authConfig
+func (am AuthenticationMiddleware) ApplyConfig(config *config.Config) {
+	*am.config = *config
 }
 
 // newAuthenticator returns the authentication mode used by the request and its credentials
-func newAuthenticator(r *http.Request, authConfig *pkg.Authn, logger *zap.Logger) (Authenticator, error) {
+func newAuthenticator(r *http.Request, config *config.Config, logger *zap.Logger) (Authenticator, error) {
 	// OAuth token is favorite authentication mode
 	token := r.Header.Get("X-Id-Token")
 	if token != "" {
-		logger.Debug(fmt.Sprintf("OAuth Token = %s", token))
 		return OAuthAuthenticator{
-			token:      token,
-			authConfig: authConfig,
-			logger:     logger,
+			token:  token,
+			config: config,
+			logger: logger,
 		}, nil
 	}
 	// If no oauth token, we are looking for basicAuth
 	user, pwd, ok := r.BasicAuth()
 	if ok {
 		return BasicAuthenticator{
-			user:       user,
-			pwd:        pwd,
-			authConfig: authConfig,
-			logger:     logger,
+			user:   user,
+			pwd:    pwd,
+			config: config,
+			logger: logger,
 		}, nil
 	}
 	return nil, errors.New("unsupported authentication")