+++ internal/app/grafana-multi-tenant-proxy/server.go
@@ -11,16 +11,13 @@ import (
 	"github.com/prometheus/client_golang/prometheus/collectors"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-
-	"github.com/giantswarm/grafana-multi-tenant-proxy/internal/app/grafana-multi-tenant-proxy/auth"
-	"github.com/giantswarm/grafana-multi-tenant-proxy/internal/pkg"
 	"github.com/urfave/cli/v2"
 	"go.uber.org/zap"
-)
 
-var keepOrgID bool
-var authConfigLocation string
-var authConfig *pkg.Authn
+	"github.com/giantswarm/grafana-multi-tenant-proxy/internal/app/grafana-multi-tenant-proxy/config"
+	"github.com/giantswarm/grafana-multi-tenant-proxy/internal/app/grafana-multi-tenant-proxy/handler"
+	"github.com/giantswarm/grafana-multi-tenant-proxy/internal/app/grafana-multi-tenant-proxy/handler/auth"
+)
 
 var (
 	requestsTotal = promauto.NewCounterVec(prometheus.CounterOpts{
@@ -46,73 +43,46 @@ var (
 	)
 )
 
-func loadConfig() (*pkg.Authn, error) {
-	config, err := pkg.ParseConfig(&authConfigLocation)
-	config.KeepOrgID = keepOrgID
-	return config, err
+func initLogger(logLevel string) (*zap.Logger, error) {
+	zapConfig := zap.NewProductionConfig()
+	level, err := zap.ParseAtomicLevel(logLevel)
+	if err != nil {
+		return nil, err
+	}
+	zapConfig.Level = level
+  return zapConfig.Build()
 }
 
-// Serve serves
+// Serve serves requests to the proxy
 func Serve(c *cli.Context) error {
-	targetServerURL, _ := url.Parse(c.String("target-server"))
-	addr := fmt.Sprintf(":%d", c.Int("port"))
-	authConfigLocation = c.String("auth-config")
-	keepOrgID = c.Bool("keep-orgid")
 	logLevel := c.String("log-level")
 	if logLevel == "" {
 		logLevel = "INFO"
 	}
 
-	var logger *zap.Logger
-	{
-		zapConfig := zap.NewProductionConfig()
-		level, err := zap.ParseAtomicLevel(logLevel)
-		if err != nil {
-			return cli.Exit(fmt.Sprintf("Could not parse log level %v", err), -1)
-		}
-		zapConfig.Level = level
-
-		logger = zap.Must(zapConfig.Build())
-		defer logger.Sync()
+	logger, err := initLogger(logLevel)
+	if err != nil {
+		return cli.Exit(fmt.Sprintf("Could not create logger %v", err), -1)
 	}
+	// Ensure that the logger is flushed before the program exits
+	defer logger.Sync()
 
 	errorLogger, err := zap.NewStdLogAt(logger, zap.ErrorLevel)
 	if err != nil {
-		return cli.Exit(fmt.Sprintf("Could not create standard logger %v", err), -1)
+		return cli.Exit(fmt.Sprintf("Could not create standard error logger %v", err), -1)
 	}
 
-	authConfig, err = loadConfig()
+	// Read the configuration
+	proxyConfigLocation := c.String("proxy-config")
+	authConfigLocation := c.String("auth-config")
+	cfg, err := config.ReadConfigFiles(proxyConfigLocation, authConfigLocation)
 	if err != nil {
 		return cli.Exit(fmt.Sprintf("Could not parse config %v", err), -1)
 	}
 
-	var reverseProxy *httputil.ReverseProxy
-	{
-		reverseProxy = &httputil.ReverseProxy{
-			Rewrite: func(r *httputil.ProxyRequest) {
-				r.SetURL(targetServerURL)
-				r.Out.Host = targetServerURL.Host
-				r.Out.Header.Set("X-Forwarded-Host", targetServerURL.Host)
-				orgID := r.In.Context().Value(auth.OrgIDKey)
-
-				if orgID != "" {
-					r.Out.Header.Set("X-Scope-OrgID", orgID.(string))
-				}
-			},
-			ErrorLog: errorLogger,
-		}
-	}
-
-	authenticationMiddleware := auth.NewAuthenticationMiddleware(
-		logger,
-		ReverseTarget(reverseProxy),
-		*authConfig,
-	)
-
-	handlers := Logger(
-		authenticationMiddleware.Authenticate(),
-		logger,
-	)
+	proxy := handler.NewProxy(cfg, logger, errorLogger)
+	authenticationMiddleware := auth.NewAuthenticationMiddleware(cfg, logger, proxy.Handler())
+	handlers := handler.Logger(authenticationMiddleware.Authenticate(), logger)
 
 	// Register Prometheus collectors
 	prometheus.MustRegister(collectors.NewBuildInfoCollector())
@@ -120,23 +90,28 @@ func Serve(c *cli.Context) error {
 	// We handle metrics first to avoid calling the authentication middleware
 	http.Handle("/metrics", promhttp.Handler())
 	http.HandleFunc("/", instrumentHandler("default", handlers))
+
+	// Reload config endpoint
 	http.HandleFunc("/-/reload", func(w http.ResponseWriter, r *http.Request) {
 		if r.Method != "POST" {
 			http.Error(w, "Invalid request method.", http.StatusMethodNotAllowed)
 			return
 		}
-		authConfig, err := loadConfig()
 
+		cfg, err = config.ReadConfigFiles(proxyConfigLocation, authConfigLocation)
 		if err != nil {
 			logger.Error("Could not reload config", zap.Error(err))
 			w.WriteHeader(500)
 		} else {
-			authenticationMiddleware.ApplyConfig(*authConfig)
+			authenticationMiddleware.ApplyConfig(cfg)
+			proxy.ApplyConfig(cfg)
 			w.WriteHeader(200)
 			w.Write([]byte("OK"))
 		}
 	})
 
+	// Start the server
+	addr := fmt.Sprintf(":%d", c.Int("port"))
 	server := &http.Server{Addr: addr, ErrorLog: errorLogger}
 	if err := server.ListenAndServe(); err != nil {
 		return cli.Exit(fmt.Sprintf("Grafana multi tenant proxy could not start %v", err), -1)